requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
cffi==2.0.0
discord.py==2.6.3
frozenlist==1.8.0
idna==3.10
multidict==6.7.0
propcache==0.4.0
pycparser==2.23
PyNaCl==1.6.0
python-dotenv==1.1.1
PyYAML==6.0.3
yarl==1.22.0
